+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### New Features
 
+* Parsing `Header and Footer` in xlsx
+
+### New Features
+
 * Parsing `Defined Names` in xlsx
 
 ## 0.13.0 (2021-05-25)